+++ Tests/WorkflowTest.php
@@ -15,6 +15,7 @@ use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\TestWith;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\EventDispatcher\EventDispatcher;
+use Symfony\Component\Workflow\Arc;
 use Symfony\Component\Workflow\Definition;
 use Symfony\Component\Workflow\Event\EnteredEvent;
 use Symfony\Component\Workflow\Event\Event;
@@ -867,6 +868,133 @@ class WorkflowTest extends TestCase
             'b' => 3,
             'c' => 1,
         ], $marking);
+    }
+
+    public function testWithArcAndWeight()
+    {
+        //              ┌───────────────────┐     ┌─────────────┐     ┌─────────────┐  4
+        //              │    prepare_leg    │ ──▶ │  build_leg  │ ──▶ │ leg_created │ ───────────────────────────┐
+        //              └───────────────────┘     └─────────────┘     └─────────────┘                            │
+        //                ▲                                                                                      │
+        //                │ 4                                                                                    │
+        //                │                                                                                      ▼
+        // ┌──────┐     ┌───────────────────┐     ┌─────────────┐     ┌─────────────┐      ┌─────────────┐     ┌──────┐     ┌──────────┐
+        // │ init │ ──▶ │       start       │ ──▶ │ prepare_top │ ──▶ │  build_top  │ ───▶ │ top_created │ ──▶ │ join │ ──▶ │ finished │
+        // └──────┘     └───────────────────┘     └─────────────┘     └─────────────┘      └─────────────┘     └──────┘     └──────────┘
+        //                │                                                                                      ▲
+        //                │                                                                                      │
+        //                ▼                                                                                      │
+        //              ┌───────────────────┐                                                                    │
+        //              │ stopwatch_running │ ───────────────────────────────────────────────────────────────────┘
+        //              └───────────────────┘
+        //
+        // make_table:
+        //     transitions:
+        //         start:
+        //             from: init
+        //             to:
+        //                 -   place: prepare_leg
+        //                     weight: 4
+        //                 -   place: prepare_top
+        //                     weight: 1
+        //                 -   place: stopwatch_running
+        //                     weight: 1
+        //         build_leg:
+        //             from: prepare_leg
+        //             to: leg_created
+        //         build_top:
+        //             from: prepare_top
+        //             to: top_created
+        //         join:
+        //             from:
+        //                 - place: leg_created
+        //                   weight: 4
+        //                 - top_created
+        //                 - stopwatch_running
+        //             to: finished
+
+        $definition = new Definition(
+            [],
+            [
+                new Transition('start', 'init', [new Arc('prepare_leg', 4), 'prepare_top', 'stopwatch_running']),
+                new Transition('build_leg', 'prepare_leg', 'leg_created'),
+                new Transition('build_top', 'prepare_top', 'top_created'),
+                new Transition('join', [new Arc('leg_created', 4), 'top_created', 'stopwatch_running'], 'finished'),
+            ]
+        );
+
+        $subject = new Subject();
+        $workflow = new Workflow($definition);
+
+        $this->assertTrue($workflow->can($subject, 'start'));
+        $this->assertFalse($workflow->can($subject, 'build_leg'));
+        $this->assertFalse($workflow->can($subject, 'build_top'));
+        $this->assertFalse($workflow->can($subject, 'join'));
+
+        $workflow->apply($subject, 'start');
+
+        $this->assertSame([
+            'prepare_leg' => 4,
+            'prepare_top' => 1,
+            'stopwatch_running' => 1,
+        ], $subject->getMarking());
+        $this->assertTrue($workflow->can($subject, 'build_leg'));
+        $this->assertTrue($workflow->can($subject, 'build_top'));
+        $this->assertFalse($workflow->can($subject, 'join'));
+
+        $workflow->apply($subject, 'build_leg');
+
+        $this->assertSame([
+            'prepare_leg' => 3,
+            'prepare_top' => 1,
+            'stopwatch_running' => 1,
+            'leg_created' => 1,
+        ], $subject->getMarking());
+        $this->assertTrue($workflow->can($subject, 'build_leg'));
+        $this->assertTrue($workflow->can($subject, 'build_top'));
+        $this->assertFalse($workflow->can($subject, 'join'));
+
+        $workflow->apply($subject, 'build_top');
+
+        $this->assertSame([
+            'prepare_leg' => 3,
+            'stopwatch_running' => 1,
+            'leg_created' => 1,
+            'top_created' => 1,
+        ], $subject->getMarking());
+        $this->assertTrue($workflow->can($subject, 'build_leg'));
+        $this->assertFalse($workflow->can($subject, 'build_top'));
+        $this->assertFalse($workflow->can($subject, 'join'));
+
+        $workflow->apply($subject, 'build_leg');
+
+        $this->assertSame([
+            'prepare_leg' => 2,
+            'stopwatch_running' => 1,
+            'leg_created' => 2,
+            'top_created' => 1,
+        ], $subject->getMarking());
+        $this->assertTrue($workflow->can($subject, 'build_leg'));
+        $this->assertFalse($workflow->can($subject, 'build_top'));
+        $this->assertFalse($workflow->can($subject, 'join'));
+
+        $workflow->apply($subject, 'build_leg');
+        $workflow->apply($subject, 'build_leg');
+
+        $this->assertSame([
+            'stopwatch_running' => 1,
+            'leg_created' => 4,
+            'top_created' => 1,
+        ], $subject->getMarking());
+        $this->assertFalse($workflow->can($subject, 'build_leg'));
+        $this->assertFalse($workflow->can($subject, 'build_top'));
+        $this->assertTrue($workflow->can($subject, 'join'));
+
+        $workflow->apply($subject, 'join');
+
+        $this->assertSame([
+            'finished' => 1,
+        ], $subject->getMarking());
     }
 
     private function assertPlaces(array $expected, Marking $marking)